# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/customer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Management"
                ],
                "summary": "Order credit notation",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CustomerPostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CustomerCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errormodel.BusinessError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errormodel.BusinessError"
                        }
                    }
                }
            }
        },
        "/v1/dog/{breed}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Management"
                ],
                "summary": "Order credit notation",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DogPostRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errormodel.BusinessError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errormodel.BusinessError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "enum.DogBreed": {
            "type": "string",
            "enum": [
                "CHOW",
                "HOUND"
            ],
            "x-enum-varnames": [
                "Chow",
                "Hound"
            ]
        },
        "errormodel.BusinessError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "request.CustomerPostRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "request.DogPostRequest": {
            "type": "object",
            "properties": {
                "breed": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.DogBreed"
                        }
                    ],
                    "example": "hound"
                }
            }
        },
        "response.CustomerCreateResponse": {
            "type": "object",
            "properties": {
                "customerID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "JWT": {
            "description": "JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}